backend/requirements.txt
-i https://pypi.org/simple
asgiref==3.9.0; python_version >= '3.9'
attrs==25.3.0; python_version >= '3.8'
autobahn==24.4.2; python_version >= '3.9'
automat==25.4.16; python_version >= '3.9'
cffi==1.17.1; python_version >= '3.8'
channels[daphne]==4.2.2; python_version >= '3.8'
constantly==23.10.4; python_version >= '3.8'
cryptography==45.0.5; python_version >= '3.7' and python_full_version not in '3.9.0, 3.9.1'
daphne==4.2.1; python_version >= '3.9'
django==5.2.4; python_version >= '3.10'
django-auto-prefetching==0.2.12
django-cors-headers==4.0.0
django-filter==23.2
djangorestframework==3.16.0; python_version >= '3.9'
djangorestframework-simplejwt==5.5.0; python_version >= '3.9'
hyperlink==21.0.0
idna==3.10; python_version >= '3.6'
incremental==24.7.2; python_version >= '3.8'
markdown==3.4.3
psycopg2-binary==2.9.6
pyasn1==0.6.1; python_version >= '3.8'
pyasn1-modules==0.4.2; python_version >= '3.8'
pycparser==2.22; python_version >= '3.8'
pyjwt==2.9.0; python_version >= '3.8'
pyopenssl==25.1.0; python_version >= '3.7'
pytz==2025.2
service-identity==24.2.0; python_version >= '3.8'
setuptools==80.9.0; python_version >= '3.9'
sqlparse==0.5.3; python_version >= '3.8'
twisted[tls]==25.5.0; python_full_version >= '3.8.0'
txaio==25.6.1; python_version >= '3.10'
typing-extensions==4.14.0; python_version >= '3.9'
tzdata==2025.2; python_version >= '2'
zope-interface==7.2; python_version >= '3.8'
